+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Support for Hydreon RG-15 Solid State Rain sensor (#12974)
 - Support for IKEA VINDRIKTNING particle concentration sensor (#12976)
 - Berry support for Curve 25519 EC crypto
+- Command ``SetOption2 1`` to enable display of global temperature/humidity/pressure info to JSON sensor message
 
 ### Changed
 - Shelly EM template needs to use GPIO ADE7953_IRQ_2